restructure: weather widget utilities; optimize weather service logic
