+++ utils/weather/types.ts
@@ -1,0 +1,22 @@
+export interface WeatherCondition {
+  FeelsLikeC: string;
+  FeelsLikeF: string;
+  humidity: string;
+  tempC: string; // normalized
+  tempF: string;
+  weatherDesc: { value: string }[];
+  [key: string]: any;
+}
+
+export interface WeatherForecast {
+  date: string;
+  maxtempC: string;
+  mintempC: string;
+  hourly: WeatherCondition[];
+  [key: string]: any;
+}
+
+export interface WeatherData {
+  current: WeatherCondition | null;
+  forecast: WeatherForecast[];
+}
